Use custom Outcome class instead of Kotlin Result (#579)
Unfortunately the Kotlin compiler puts hard restrictions on the Result class, such as it not being allowed to be used on the LHS of the Elvis operator. This is no good for our use case, so I've switched it out for our own custom `Outcome` class which is similar. `kotlin.Result` is "experimental" anyways.