Use the template engine to build the forms query (#590)
This will be really useful when we get to sorting and filtering of form instances. The options selected in the UI can be easily managed into a query by this style of querying.

This change also gets rid of the getByFeatureGuidAndCreatorAccountGuid method, since the getByFeatureGuid method is now flexible enough to handle both use cases.