Require explicit opt-in to use other modules (#606)
Code that interacts synchronously with other modules is an antipattern. It's not a huge deal now since the application gets run as a monolith, but eventually it will be run as microservices instead, in which case we should avoid this. This change requires all usages of service methods from other modules to have an explicit opt-in.