Disable errors (use warnings) for comments in JSON (#716)
By default, IntelliJ flags comments in JSON files as errors. This is fair because it's a violation of the JSON standard, but in reality many JSON parsers can handle comments just fine, and many JSON files aren't even intended to be parsed, but to serve as documentation. By changing this to a warning we leave it up to the developer whether to be concerned about comments in JSON files or not.