Create permissions and auth common modules (#919)
By separating permissions and auth information into separate modules, we can share those modules only as needed. This helps with both the client interaction work that is underway, as well as with getting rid of deprecated Gradle modules that currently solve this for tests by sharing every module with every other module (undesirable).