Allow creating users without names (#925)
Users can be created without names already at the service layer, but that nullability was previously not propagated to the REST layer. The use case for this is Auth0's passwordless authentication. By propagating this to the service layer we allow the JwtClaimsRequestService to interact with the user module via REST instead of at the service layer.