Create server-to-server HTTP client (#921)
The server-to-server HTTP client generates its own JWTs which will be exempt from permission checks, caching them for performance and refreshing them as necessary. It'll be used for server-to-server interaction (the monolith will now communicate with itself over REST, so that it's using Clients instead of Services for server-to-server interaction and is therefore more ready to split to microservices when necessary), but NOT for the GraphQL backend to communicate with other servers, since the GraphQL backend will forward the client's original JWT instead.