Merge forms module service interface into module (#947)
The service-interface was used when inter-module interaction occurred at the service layer. Now it occurs over REST, so the service-interface is not necessary. The same classes exist, but they are moved into the same Gradle module as the rest of the code, instead of being split into their own special module.