Get rid of user role PUT/DELETE endpoints (#974)
Instead of having separate endpoints to PUT and DELETE user roles, they can simply be included in the user PATCH endpoint. This was not done previously because users are allowed to modify their own data, but cannot add their own user roles. This is now solved by adding an additional conditional permissions check in the PATCH endpoint.